+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "vscode-pin-up" extension will be documented in this 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.2.0] - 2020-06-06
+
+- Added: Create file/folder and delete it in Pinned List.
+- Fixed: Add category property to each commands.
+
 ## [1.1.1] - 2020-05-29
 
 - Added: I18N Support of Chinese.